moved UDPMsgConn to tubes testing package
